Move : CellView 다른 파일로 분리
